Remove .flutter-plugins-dependencies
This file is auto-generated by Flutter based on the current plugin dependencies in your project.

This file is specific to your local development environment and can vary between different environments or developer machines, depending on the installed plugin versions. Including it in version control can lead to unnecessary merge conflicts and does not provide any benefits, as each developer's environment will regenerate it based on the pubspec.yaml and pubspec.lock files, which should be versioned.

issue #17