fix: Add secret management functionality
This commit introduces the env module to facilitate the concealment and management of sensitive information such as APIs, and adds unit tests to verify the effectiveness of retrieving information from env. It updates the Flutter framework to the latest version 3.19.0, which automatically updates and introduces new components required in iOS and Android files, among others.

* fix: Added dotenv related code in main.dart to ensure the program can obtain API information through env
* fix: Applied dotenv dependency in firebase_options to securely fetch API from env, instead of using it directly
* fix: Updated the Flutter framework to the latest version 3.19.0, which automatically updates, adds, and removes for the new components required in iOS, Android, etc.
* fix: Added dotenv dependency in pubspec.yaml, and included asset- env to ensure the program can find env in the root directory

addresses #17