fix: Redesigned the UI for Speech to Text and stabilized its basic functionality.
This commit redesigns the UI for Speech to Text and stabilizes its functionality by modifying UI design elements in main.dart to make the Speech to Text button more intuitive and simple, as well as adding text to display the status of voice conversion. Furthermore, this update equips the button with basic and stable functionality and includes unit test code.

*fix: Added and updated related UI code in main.dart,
*fix: Added and updated actual functional code for Speech to Text in main.dart, ensuring it works effectively on both Android and iOS
*update: Added a unit test file, speech_test.dart, in the test directory to test the functionality of Speech to Text
*update: Added dependencies for Speech to Text and necessary testing purposes, such as unit tests, in pubspec.yaml
*fix: Added microphone and speech recognition permissions to the IOS and Android permission files.

addresses #12