Feature: News Summary Functionality and Testing
This commit implements the news summary functionality by utilizing the News API. When the News Summary button is clicked, it enters the news_service to retrieve the top 3 news articles, which are then sent to GPT-3.5 for summarization and subsequently broadcasted. If the user clicks the button within an hour, the previously summarized content will be broadcasted to ensure the API is not abused.

* fix: Added _buildNewsButton function and related code in home.dart.
* fix: Added the NEWS_API_KEY API key in .env. (.env is hidden)
* fix: Modified targetSdkVersion to 28 in app/build.gradle to ensure wider compatibility with more Android devices.
* fix: Added news_service.dart in lib/backend to retrieve news content through the API.
* fix: Added news_service_test.dart in test/unit to test the normal operation of news_service.
* fix: Modified the handling of the news category in lib/backend/text_to_gpt_service.dart.

Addresses #55